Issue #11163: enforce file size on inputs for MethodCount check
